% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdiSalOrder.R
\name{EdiSalOrder_delete}
\alias{EdiSalOrder_delete}
\title{EDI销售订单删除}
\usage{
EdiSalOrder_delete(erp_token, FBillNO)
}
\arguments{
\item{FBillNO}{}
}
\value{
两个数的和
}
\description{
EDI销售订单删除
}
\examples{
EdiSalOrder_delete()
}
