% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdiSalOrder.R
\name{EdiSalOrder_view}
\alias{EdiSalOrder_view}
\title{EDI销售订单查询}
\usage{
EdiSalOrder_view(dms_token, FBillNO)
}
\arguments{
\item{FBillNO}{}
}
\value{
两个数的和
}
\description{
EDI销售订单查询
}
\examples{
EdiSalOrder_view()
}
